Got Linux build working again, added JTAG capability
This adds the ability to program the CPLD. It's a bit of a bodge, since
strictly, the JTAG pins should be manipulated at 3.3v not 5v (as this
hardware does) but it does seem to work (well, it does now I fixed a
whole load of 64-bit bugs in the ancient Altera jam player code!)

+++ libs/libz80board/src/jam_player/jamcomp.c
@@ -114,7 +114,7 @@ short jam_read_packed(char *buffer, long length, short bits)
 long jam_uncompress
 (
 	char *in, 
-	long in_length, 
+	int in_length, 
 	char *out, 
 	long out_length,
 	int version
