+++ libs/libz80board/src/jam_player/jamcomp.h
@@ -15,7 +15,7 @@
 long jam_uncompress
 (
 	char *in, 
-	long in_length, 
+	int in_length, 
 	char *out, 
 	long out_length,
 	int version
